requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pillow
